test: add web layer unit tests for RiskEvaluationController with API key security
reactivate Security Filter

+++ fdw-analytics-engine/src/main/java/com/chriscodecc/fdw_analytics_engine/dto/RollingMetricProjection.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface RollingMetricProjection {
-    Long getCompanyId();
+    Integer getCompanyId();
     String getName();
     LocalDate getFullDate();
     BigDecimal getClosePrice();
